requirements.txt
requests>=2.28.0
urllib3>=1.26.0
customtkinter>=5.0.0
pywebview>=4.0.0
Pillow>=9.0.0
rich>=13.0.0
